+++ packages/xchain-aggregator/src/protocols/mayachain/mayachainProtocol.ts
@@ -1,5 +1,5 @@
 import { AssetCacao, MAYAChain } from '@xchainjs/xchain-mayachain'
-import { MayachainAMM } from '@xchainjs/xchain-mayachain-amm'
+import { ApproveParams, IsApprovedParams, MayachainAMM } from '@xchainjs/xchain-mayachain-amm'
 import { MayachainQuery } from '@xchainjs/xchain-mayachain-query'
 import {
   AnyAsset,
@@ -10,6 +10,7 @@ import {
   isSynthAsset,
   isTradeAsset,
 } from '@xchainjs/xchain-util'
+import { Wallet } from '@xchainjs/xchain-wallet'
 
 import {
   IProtocol,
@@ -27,12 +28,36 @@ export class MayachainProtocol implements IProtocol {
   public readonly name = 'Mayachain'
   private mayachainQuery: MayachainQuery
   private mayachainAmm: MayachainAMM
-  private configuration: Omit<ProtocolConfig, 'wallet'> | undefined
+  private configuration: ProtocolConfig | undefined
+  private wallet?: Wallet
 
   constructor(configuration?: ProtocolConfig) {
     this.mayachainQuery = new MayachainQuery()
     this.mayachainAmm = new MayachainAMM(this.mayachainQuery, configuration?.wallet)
     this.configuration = configuration
+    this.wallet = configuration?.wallet
+  }
+  /**
+   * Aprove tx for ERC-20
+   * @param {ApproveParams} approveParams params to approve tx
+   * @returns {TxSubmitted} Transaction hash and URL of the swap
+   */
+  async approveRouterToSpend(params: ApproveParams): Promise<TxSubmitted> {
+    const { asset, amount } = params
+    const txSubmitted = await this.mayachainAmm.approveRouterToSpend({ asset, amount })
+    await this.wallet?.awaitTxConfirmed(asset.chain, txSubmitted.hash)
+    return txSubmitted
+  }
+
+  /**
+   * Check if tx is approved for ERC-20
+   * @param {IsApprovedParams} isApprovedParams params to check if tx is approved
+   * @returns {string[]} array of errors
+   */
+  async shouldBeApproved(params: IsApprovedParams): Promise<boolean> {
+    const { asset, amount, address } = params
+    const errors = await this.mayachainAmm.isRouterApprovedToSpend({ asset, amount, address })
+    return errors.some((error) => error === 'Maya router has not been approved to spend this amount')
   }
 
   /**