+++ examples/loans/CHANGELOG.md
@@ -1,5 +1,56 @@
 # xchainjs-loans
 
+## 1.0.24
+
+### Patch Changes
+
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+  - @xchainjs/xchain-evm@1.0.4
+  - @xchainjs/xchain-thorchain-amm@2.0.4
+  - @xchainjs/xchain-util@1.0.1
+  - @xchainjs/xchain-client@1.0.2
+  - @xchainjs/xchain-cosmos@2.0.2
+  - @xchainjs/xchain-evm-providers@1.0.2
+  - @xchainjs/xchain-midgard-query@1.0.2
+  - @xchainjs/xchain-thorchain-query@1.0.2
+  - @xchainjs/xchain-utxo-providers@1.0.2
+
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [0bf8f3e]
+  - @xchainjs/xchain-evm@1.0.3
+  - @xchainjs/xchain-thorchain-amm@2.0.3
+
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies [4d96d34]
+  - @xchainjs/xchain-evm@1.0.2
+  - @xchainjs/xchain-thorchain-amm@2.0.2
+
+## 1.0.21
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+- Updated dependencies [837e3e7]
+- Updated dependencies [323cbea]
+  - @xchainjs/xchain-thorchain-query@1.0.1
+  - @xchainjs/xchain-utxo-providers@1.0.1
+  - @xchainjs/xchain-evm-providers@1.0.1
+  - @xchainjs/xchain-midgard-query@1.0.1
+  - @xchainjs/xchain-thorchain-amm@2.0.1
+  - @xchainjs/xchain-thornode@0.3.19
+  - @xchainjs/xchain-midgard@0.5.9
+  - @xchainjs/xchain-client@1.0.1
+  - @xchainjs/xchain-cosmos@2.0.1
+  - @xchainjs/xchain-evm@1.0.1
+
 ## 1.0.20
 
 ### Patch Changes