+++ packages/xchain-aggregator/CHANGELOG.md
@@ -1,5 +1,56 @@
 # Changelog
 
+## 1.0.4
+
+### Patch Changes
+
+- b07b69a: Support auto-approval for ERC-20 tokens
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+  - @xchainjs/xchain-wallet@1.0.4
+  - @xchainjs/xchain-mayachain-amm@3.0.4
+  - @xchainjs/xchain-thorchain-amm@2.0.4
+  - @xchainjs/xchain-util@1.0.1
+  - @xchainjs/xchain-client@1.0.2
+  - @xchainjs/xchain-mayachain@2.0.2
+  - @xchainjs/xchain-mayachain-query@1.0.2
+  - @xchainjs/xchain-thorchain@2.0.2
+  - @xchainjs/xchain-thorchain-query@1.0.2
+
+## 1.0.3
+
+### Patch Changes
+
+- @xchainjs/xchain-thorchain-amm@2.0.3
+- @xchainjs/xchain-wallet@1.0.3
+- @xchainjs/xchain-mayachain-amm@3.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @xchainjs/xchain-thorchain-amm@2.0.2
+- @xchainjs/xchain-wallet@1.0.2
+- @xchainjs/xchain-mayachain-amm@3.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+- Updated dependencies [837e3e7]
+- Updated dependencies [323cbea]
+- Updated dependencies [be1c45f]
+  - @xchainjs/xchain-mayachain-query@1.0.1
+  - @xchainjs/xchain-thorchain-query@1.0.1
+  - @xchainjs/xchain-mayachain-amm@3.0.1
+  - @xchainjs/xchain-thorchain-amm@2.0.1
+  - @xchainjs/xchain-mayachain@2.0.1
+  - @xchainjs/xchain-thorchain@2.0.1
+  - @xchainjs/xchain-client@1.0.1
+  - @xchainjs/xchain-wallet@1.0.1
+
 ## 1.0.0
 
 ### Major Changes