+++ packages/xchain-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [b07b69a]
+  - @xchainjs/xchain-util@1.0.1
+
+## 1.0.1
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+- 323cbea: Add info types for assets
+
 ## 1.0.0
 
 ### Major Changes