+++ packages/xchain-cosmos/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [b07b69a]
+  - @xchainjs/xchain-util@1.0.1
+  - @xchainjs/xchain-client@1.0.2
+  - @xchainjs/xchain-cosmos-sdk@1.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+- Updated dependencies [837e3e7]
+- Updated dependencies [323cbea]
+  - @xchainjs/xchain-client@1.0.1
+  - @xchainjs/xchain-cosmos-sdk@1.0.1
+
 ## 2.0.0
 
 ### Major Changes