+++ packages/xchain-mayamidgard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.6
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+
 ## 0.1.5
 
 ### Patch Changes