+++ packages/xchain-thornode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.19
+
+### Patch Changes
+
+- 837e3e7: Axios version update to v1.7.4
+
 ## 0.3.18
 
 ### Patch Changes