+++ tools/txJammer/CHANGELOG.md
@@ -1,5 +1,72 @@
 # txJammer
 
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+- Updated dependencies [b07b69a]
+  - @xchainjs/xchain-wallet@1.0.4
+  - @xchainjs/xchain-thorchain-amm@2.0.4
+  - @xchainjs/xchain-util@1.0.1
+  - @xchainjs/xchain-avax@1.0.4
+  - @xchainjs/xchain-bsc@1.0.4
+  - @xchainjs/xchain-ethereum@1.0.4
+  - @xchainjs/xchain-binance@6.0.2
+  - @xchainjs/xchain-bitcoin@1.0.2
+  - @xchainjs/xchain-bitcoincash@1.0.2
+  - @xchainjs/xchain-client@1.0.2
+  - @xchainjs/xchain-cosmos@2.0.2
+  - @xchainjs/xchain-doge@1.0.2
+  - @xchainjs/xchain-litecoin@1.0.2
+  - @xchainjs/xchain-thorchain@2.0.2
+  - @xchainjs/xchain-thorchain-query@1.0.2
+
+## 1.0.17
+
+### Patch Changes
+
+- @xchainjs/xchain-avax@1.0.3
+- @xchainjs/xchain-bsc@1.0.3
+- @xchainjs/xchain-ethereum@1.0.3
+- @xchainjs/xchain-thorchain-amm@2.0.3
+- @xchainjs/xchain-wallet@1.0.3
+
+## 1.0.16
+
+### Patch Changes
+
+- @xchainjs/xchain-avax@1.0.2
+- @xchainjs/xchain-bsc@1.0.2
+- @xchainjs/xchain-ethereum@1.0.2
+- @xchainjs/xchain-thorchain-amm@2.0.2
+- @xchainjs/xchain-wallet@1.0.2
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [837e3e7]
+- Updated dependencies [323cbea]
+- Updated dependencies [be1c45f]
+  - @xchainjs/xchain-thorchain-query@1.0.1
+  - @xchainjs/xchain-thorchain-amm@2.0.1
+  - @xchainjs/xchain-bitcoincash@1.0.1
+  - @xchainjs/xchain-thorchain@2.0.1
+  - @xchainjs/xchain-litecoin@1.0.1
+  - @xchainjs/xchain-thornode@0.3.19
+  - @xchainjs/xchain-binance@6.0.1
+  - @xchainjs/xchain-bitcoin@1.0.1
+  - @xchainjs/xchain-midgard@0.5.9
+  - @xchainjs/xchain-client@1.0.1
+  - @xchainjs/xchain-cosmos@2.0.1
+  - @xchainjs/xchain-doge@1.0.1
+  - @xchainjs/xchain-wallet@1.0.1
+  - @xchainjs/xchain-avax@1.0.1
+  - @xchainjs/xchain-bsc@1.0.1
+  - @xchainjs/xchain-ethereum@1.0.1
+
 ## 1.0.14
 
 ### Patch Changes